Move non-test helpers in helpers.go
